pyproject.toml
[project]
name = "weather-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.58.2",
    "azure-identity>=1.23.1",
    "azure-keyvault-secrets>=4.10.0",
    "mcp>=1.12.0",
    "openai>=1.97.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
]
